Make Avatar own component
