+++ src/components/Header/Header.css
@@ -54,18 +54,6 @@ div.HeaderLinksWrap.LoggedIn .BelowHeader {
   top: 100%;
 }
 
-.UserAvatar {
-  width: 3rem;
-  height: 3rem;
-  border-radius: 50%;
-  text-align: center;
-  font-weight: bold;
-  font-size: 24px;
-  display: flex;
-  align-items: center; /* horizontal */
-  justify-content: center; /* vertical */
-}
-
 .DropdownArrowSvg {
   width: 2rem;
   height: 2rem;
